Feat: 모델 파트 구현
